requirements.txt
anyio==3.3.2
asgiref==3.4.1
autopep8==1.5.7
cffi==1.15.0
click==8.0.1
cryptography==35.0.0
dnspython==1.16.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
fastapi-login==1.7.2
fastapi_permissions==0.2.7
gunicorn==20.1.0
h11==0.12.0
idna==3.2
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pymongo==3.12.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
